remove_dumps - only bitstream names which begin with "db_dump_" will be deleted to allow to store other data than dumps (persistent snapshots)

+++ appconfig/tasks/other.py
@@ -57,10 +57,12 @@ def list_dumps(app):
 @task_app_from_environment
 def remove_dumps(app, keep=10):
     if app.dbdump:
-        for i, bs in enumerate(cdstar.get_bitstreams(app.dbdump), start=1):
-            if i > int(keep):
+        i = 0
+        for bs in cdstar.get_bitstreams(app.dbdump):
+            if i > int(keep) and bs.name.startswith('db_dump_'):
                 print('deleting dump {0}'.format(bs.name))
                 bs.bitstream.delete()
+                i += 1
 
 
 @task_app_from_environment
